requirements.txt
numpy
pillow
torch
opencv-python
numpy
tqdm
pickle
torch
deepface
